requirements.txt
redbook-cli>=0.1.0
playwright>=1.40.0
pyyaml>=6.0
Pillow>=10.0.0
requests>=2.28.0
websockets>=11.0
